==> notebooks/3_Topic_Model_Training.ipynb <==
# Automatic Learning of Key Phrases and Topics in Document Collections

## Part 3: Topic Modeling Training and Summarization

### Overview

This notebook is Part 3 of 6, in a series providing a step-by-step description of how to process and analyze the contents of a large collection of text documents in an unsupervised manner. Using Python packages and custom code examples, we have implemented the basic framework that combines key phrase learning and latent topic modeling as described in the paper entitled ["Modeling Multiword Phrases with Constrained Phrases Tree for Improved Topic Modeling of Conversational Speech"](http://people.csail.mit.edu/hazen/publications/Hazen-SLT-2012.pdf) which was originally presented in the 2012 IEEE Workshop on Spoken Language Technology.

Although the paper examines the use of the technology for analyzing human-to-human conversations, the techniques are quite general and can be applied to a wide range natural language data including news stories, legal documents, research publications, social media forum discussion, customer feedback forms, product reviews, and many more.

Part 3 of the series shows how to train a topic model on a collection of text documents and how to use the topic model to summarize the contents of the corpus. The training is applied to text generated from the preprocessing and phrase learning stages presented in Parts 1 and 2.  


### Import Relevant Python Packages

Most significantly, Part 3 relies on the use of the [Gensim Python library](http://radimrehurek.com/gensim/)  for generating a sparse bag-of-words representation of each document and then training a [Latent Dirichlet Allocation (LDA)](https://en.wikipedia.org/wiki/Latent_Dirichlet_allocation) model on the data. LDA produces a collection of latent topics learned in a completely unsupervised fashion from the text data. Each document can then be represented with a distribution of the learned topics.

import numpy
import pandas 
import re
import math
import warnings
warnings.filterwarnings(action='ignore', category=UserWarning, module='gensim')
import gensim
from gensim import corpora
from gensim import models
from operator import itemgetter
from collections import namedtuple
import time
import gc
import sys
import multiprocessing
from __future__ import print_function

### Load Text Data

# Load full TSV file including a column of text
frame = pandas.read_csv("../Data/CongressionalDocsProcessed.tsv", sep='\t', encoding='ISO-8859-1')

print ("Total docs in corpus: %d\n" % len(frame))

# Show the first five rows of the data in the frame
frame[0:5]

### Load the Stop Word Lists
Latent topic models attempt to restrict the topic learning processing to the use of only content bearing words by excluding non-content bearing <b><i>stop words</i></b>. Manually crafted stop word lists are typically manually crafted and include common functional words such as articles, conjunctions, prepositions, pronouns, etc.

# Define a function for loading lists into dictionary hash tables
def LoadListAsHash(filename):
    listHash = {}
    fp = open(filename, encoding='utf-8')

    # Read in lines one by one stripping away extra spaces, 
    # leading spaces, and trailing spaces and inserting each
    # cleaned up line into a hash table
    re1 = re.compile(' +')
    re2 = re.compile('^ +| +$')
    for stringIn in fp.readlines():
        term = re2.sub("",re1.sub(" ",stringIn.strip('\n')))
        if term != '':
            listHash[term] = 1

    fp.close()
    return listHash 

# Load the stop-list of non-content bearing function words
stopwordHash = LoadListAsHash("../Data/function_words.txt")

# Additional words can also be manually added to the stop word list as needed
stopwordHash["foo"] = 1

### Load the Mapping of Lower-Cased Vocabulary Items to Their Most Common Surface Form

# Load surface form mappings here
fp = open("../Data/Vocab2SurfaceFormMapping.tsv", encoding='utf-8')

vocabToSurfaceFormHash = {}

# Each line in the file has two tab separated fields;
# the first is the vocabulary item used during modeling
# and the second is its most common surface form in the 
# original data
for stringIn in fp.readlines():
    fields = stringIn.strip().split("\t")
    if len(fields) != 2:
        print ("Warning: Bad line in surface form mapping file: %s" % stringIn)
    elif fields[0] == "" or fields[1] == "":
        print ("Warning: Bad line in surface form mapping file: %s" % stringIn)
    else:
        vocabToSurfaceFormHash[fields[0]] = fields[1]
fp.close()


### Do Topic Modeling on Corpus using Latent Dirichlet Allocation (LDA)

#### Create the Vocabulary Used for Topic Modeling

def CreateVocabForTopicModeling(textData,stopwordHash):

    print ("Counting words")
    numDocs = len(textData) 
    globalWordCountHash = {} 
    globalDocCountHash = {} 
    for textLine in textData:
        docWordCountHash = {}
        for word in str(textLine).split():
            if word in globalWordCountHash:
                globalWordCountHash[word] += 1
            else:
                globalWordCountHash[word] = 1
            if word not in docWordCountHash: 
                docWordCountHash[word] = 1
                if word in globalDocCountHash:
                    globalDocCountHash[word] += 1
                else:
                    globalDocCountHash[word] = 1

    minWordCount = 5;
    minDocCount = 2;
    maxDocFreq = .25;
    vocabCount = 0;
    vocabHash = {}

    excStopword = 0
    excNonalphabetic = 0
    excMinwordcount = 0
    excNotindochash = 0
    excMindoccount = 0
    excMaxdocfreq =0

    print ("Building vocab")
    for word in globalWordCountHash.keys():
        # Test vocabulary exclusion criteria for each word
        if ( word in stopwordHash ):
            excStopword += 1
        elif ( not re.search(r'[a-zA-Z]', word, 0) ):
            excNonalphabetic += 1
        elif ( globalWordCountHash[word] < minWordCount ):
            excMinwordcount += 1
        elif ( word not in globalDocCountHash ):
            print ("Warning: Word '%s' not in doc count hash") % (word)
            excNotindochash += 1
        elif ( globalDocCountHash[word] < minDocCount ):
            excMindoccount += 1
        elif ( float(globalDocCountHash[word])/float(numDocs) > maxDocFreq ):
            excMaxdocfreq += 1
        else:
            # Add word to vocab
            vocabHash[word]= globalWordCountHash[word];
            vocabCount += 1 
    print ("Excluded %d stop words" % (excStopword))       
    print ("Excluded %d non-alphabetic words" % (excNonalphabetic))  
    print ("Excluded %d words below word count threshold" % (excMinwordcount)) 
    print ("Excluded %d words below doc count threshold" % (excMindoccount))
    print ("Excluded %d words above max doc frequency" % (excMaxdocfreq)) 
    print ("Final Vocab Size: %d words" % vocabCount)
            
    return vocabHash
                    
vocabHash = CreateVocabForTopicModeling(frame['ProcessedText'],stopwordHash)

# Show that the stop word "and" is not the vocabulary
'and' in vocabHash

# Show a learned phrase is in the vocabulary
'department_of_labor' in vocabHash

# The vocabulary hash table contains the total count of the vocabulary item in the data set
vocabHash["department_of_labor"]

# Print the 10 most frequent non-excluded words in the vocabulary
sorted(vocabHash.items(), key=lambda x: -x[1])[0:10]

#### Convert the Text Data Into a Sparse Vector Format

# Start by tokenizing the full text string string for each document into list of tokens
# Any token that is in not in the pre-defined set of acceptable vocabulary words is execluded
def TokenizeText(textData,vocabHash):
    tokenizedText = []
        
    for textLine in textData:
        tokenizedText.append([token for token in str(textLine).split() if token in vocabHash])    
    return tokenizedText
    
tokenizedDocs = TokenizeText(frame['ProcessedText'], vocabHash)

# Examine the tokenizaton of the first two documents
tokenizedDocs[0:2]

# Count the total number of vocabulary tokens used over the entire corpus 
numTokens = 0
for i in range(0,len(tokenizedDocs)):
    numTokens += len(tokenizedDocs[i])
print("Total number of retained tokens: %d" % numTokens)

# Create a dictionary mapping string tokens in the text to unique token IDs
dictionary = corpora.Dictionary(tokenizedDocs)

# If the reverse mapping for token ids back to string doesn't exist then create the mapping
# in the form of a list where the list index is the tokenID and the list value is the token
if len(dictionary.id2token) == 0:
    numTokens = len(dictionary.token2id);
    id2token = numTokens * [""];
    for token in dictionary.token2id:
        tokenID = dictionary.token2id[token]
        if tokenID < numTokens:
            id2token[tokenID] = token
        else: 
            print ("Warning: token id %d for token '%s' exceeds max index of %d" % (tokenID,token,numTopics-1))
    for i in range(0,numTokens):
        if id2token[i] == "":
            print ("Warning: token id %d has an empty token" % i)
else:
    id2token = dictionary.id2token
    

# The mapping from unique token ids to strings uses the id2token element of the dictionary
for i in range(0,10):
    print ("Token ID %d --> %s" % (i, id2token[i]))

# The mapping from strings to unique token ids uses the token2id element of the dictionary   
print ("")
print ("%s --> Token ID %d" % ('spent', dictionary.token2id['spent']))

# Create a Gensim corpus structure
corpus =[dictionary.doc2bow(tokens) for tokens in tokenizedDocs]

# Show that the corpus structure models the tokenized text as a sparse list of the tokens 
# in the document where each list item is represented by the unique ID for the token along 
# with the count of how often that token appeared in the document
print(tokenizedDocs[1])
print("spent:", dictionary.token2id['spent'])
print (corpus[1])

#### Train an LDA Topic Model Using the Gensim Package 

numCPUs = multiprocessing.cpu_count()
print ("Total number of CPUs:", numCPUs)

# Set the number of topics to be learned to 200
numTopics=200
if numCPUs > 1:
    numWorkers = numCPUs - 1
else:
    numWorkers = 1
numIterations = 2000

# Train LDA model 

# If you want to train a new LDA model from scratch then set this if statement to 'True'

if True:
    # If we only have one core available use standard training
    if numWorkers == 1:
        lda = gensim.models.ldamodel.LdaModel(corpus, 
                                              id2word=dictionary, 
                                              num_topics=numTopics, 
                                              iterations=numIterations,
                                              passes=10,
                                              chunksize=1000,
                                              random_state=1,
                                              offset=1024.0)
   
    # If we have multiple cores available, then use distributed multi-core training
    if numWorkers > 1:
        lda = gensim.models.ldamulticore.LdaMulticore(corpus, 
                                                      id2word=dictionary, 
                                                      num_topics=numTopics,
                                                      iterations=numIterations,
                                                      workers = numWorkers,
                                                      passes=10,
                                                      chunksize=1000,
                                                      random_state=1,
                                                      offset=1024.0)


# Saving and loading a trained LDA model
ldaFile = "../Data/CongressionalDocsLDA.pickle"

# To save a newly trained model to file set this to 'True'
if True:
    #Save a trained LDA model
    lda.save(ldaFile)

# This will load a pre-existing LDA model from file...set it to 'False' to use a newly trained model instead   
if False:
    # Loaded trained model
    lda = gensim.models.ldamodel.LdaModel.load(ldaFile)

#### Accessing the Contents of the LDA Model

# To see the accessible variables in the LDA model structure
lda.__dict__

To print out the internal help document for the LDA model class you can use the help() function.

# To print the help document for the LDA model class
help(lda)


The print_topics(n) method of the LDA model object prints out a random sampling of n different learned topics as represented by the most likely terms in the topic's langauge model, i.e. the terms that maximize the topic language model P(term|topic).

lda.print_topics(10)

#### Infer the Document Probability Score P(topic|doc) using the LDA Model

In this section, each document from the corpus is passed into the LDA model which then infers the topic distribution for each document. The topic distributions are collected into a single numpy array.

# To retrieve all topics and their probabilities we must set the LDA minimum probability setting to zero
lda.minimum_probability = 0

# This function generates the topic probabilities for each doc from the trained LDA model
# The probabilities are placed in a single matrix where the rows are documents and columns are topics
def ExtractDocTopicProbsMatrix(corpus,lda):
    # Initialize the matrix
    docTopicProbs = numpy.zeros((len(corpus),lda.num_topics))
    for docID in range(0,len(corpus)):
        for topicProb in lda[corpus[docID]]:
            docTopicProbs[docID,topicProb[0]]=topicProb[1]
    return docTopicProbs    

# docTopicProbs[docID,TopicID] --> P(topic|doc)
docTopicProbs = ExtractDocTopicProbsMatrix(corpus, lda)

# To save the document topic probabilities to file set this to True:
if True:
    docTopicProbsFile = "../Data/CongressionalDocTopicProbs.npy"
    numpy.save(docTopicProbsFile,docTopicProbs)
